% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a9a}
\alias{a9a}
\title{Adult Data Set (a9a)}
\format{A list of training and set sets representing 32561 observations,
each containing
\describe{
\item{x}{a sparse feature matrix of class \code{'dgCMatrix'} with
123 variables}
\item{y}{a binary response factor in \code{{-1, 1}}, where \code{1} indicates
a yearly income exceeding $50,000}
}}
\source{
Donors: Ronny Kohavi and Barry Becker (Data Mining and
Visualization, Silicon Graphics)

\url{https://archive.ics.uci.edu/ml/datasets/adult}

\url{https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html#a9a}
}
\usage{
a9a
}
\description{
Census data of adults.
}
\section{Preprocessing}{

From the LIBSVM website:
"The original Adult data set has 14 features, among which six are continuous
and eight are categorical. In this data set, continuous features are
discretized into quantiles, and each quantile is represented by a binary
feature. Also, a categorical feature with m categories is converted to m
binary features. Details on how each feature is converted can be found in the
beginning of each file from."
}

\keyword{binary}
