% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ijcnn1}
\alias{ijcnn1}
\title{ijcnn1}
\format{A list of training, test, and validation data, each
containing
\describe{
\item{x}{a sparse feature matrix of class \code{'dgCMatrix'} with
23 variables representing features of a a physical system}
\item{y}{a binary response factor vector in \code{{-1, 1}},
\code{-1} indicating a normal fire and \code{1} indicating
misfire}
}
with 35000, 91701, and 14990 observations respectively.}
\source{
\url{https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html#ijcnn1}

Danil Prokhorov. IJCNN 2001 neural network competition. Slide
presentation in IJCNN'01, Ford Research Laboratory, 2001.
\url{http://www.geocities.ws/ijcnn/nnc_ijcnn01.pdf}
}
\usage{
ijcnn1
}
\description{
Training, test, and validation data from the
IJCNN 2001 Neural Network Competition.
}
\section{Preprocessing}{

The dataset has been preprocessed with Winner's transformation.
}

\keyword{binary}
