% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libsvmdata.R
\docType{package}
\name{libsvmdata-package}
\alias{libsvmdata}
\alias{libsvmdata-package}
\title{libsvmdata: Data Sets from 'LIBSVM', Formatted for R}
\description{
A set of data sets from the LIBSVM data set collection, formatted
for use in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/libsvmdata}
  \item Report bugs at \url{https://github.com/jolars/libsvmdata/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (0000-0002-4029-5945)

}
\keyword{internal}
