% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phishing}
\alias{phishing}
\title{phishing}
\format{The data set is given as a list, featuring
\describe{
\item{x}{a sparse feature matrix of class \code{'dgCMatrix'} with
69 variables representing web page attributes}
\item{y}{a binary response factor vector in \code{{-1, 1}}, with
\code{-1} indicating a non-phishing website, and \code{1} indicating
the opposite}
}}
\source{
\url{https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/binary.html#phishing}

Donors: Rami Mustafa A Mohammad, Lee McCluskey, Fadi Thabtah

\url{https://archive.ics.uci.edu/ml/datasets/phishing+websites}

Yuchin Juan, Yong Zhuang, Wei-Sheng Chin, and Chih-Jen Lin.
Field-aware factorization machines for CTR prediction. In Proceedings of the
ACM Recommender Systems Conference (RecSys), 2016.
}
\usage{
phishing
}
\description{
A data set of phishing and non-phishing websites.
}
\keyword{binary}
